refactor: TripDetail 관련 패키지 이동
- uiState, uiEffect, WebView 관련
