+++ android/app/proguard-rules.pro
@@ -65,11 +65,11 @@
 -dontwarn com.google.firebase.**
 
 # WebView JavaScript Interface
--keepclassmembers class com.on.turip.ui.trip.detail.webview.WebViewVideoBridge {
+-keepclassmembers class com.on.turip.ui.compose.trip.webview.WebViewVideoBridge {
     @android.webkit.JavascriptInterface <methods>;
 }
 
--keep class com.on.turip.ui.trip.detail.webview.WebViewVideoBridge {
+-keep class com.on.turip.ui.compose.trip.webview.WebViewVideoBridge {
     public *;
 }
