+++ android/app/src/main/java/com/on/turip/ui/compose/trip/webview/TuripWebChromeClient.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.trip.detail.webview
+package com.on.turip.ui.compose.trip.webview
 
 import android.view.View
 import android.webkit.WebChromeClient
